Split classes up
